acabado, repasar

+++ juegoItems/src/Casco.java
@@ -1,5 +1,5 @@
 public class Casco extends Item{
-    public Casco(String nombre, double ataque) {
+    public Casco(String nombre, int ataque) {
         super(nombre, ataque);
     }
 }
